# Replaced ioutil.ReadAll by bufio