I did some hyperfine tests which show performance benefits.

with changes:

`hyperfine --warmup 30 'cat ~/urls | ./fff -o ./mydir'
Benchmark 1: cat ~/urls | ./fff -o ./mydir
  Time (mean ± σ):      1.988 s ±  0.013 s    [User: 0.042 s, System: 0.032 s]
  Range (min … max):    1.972 s …  2.015 s    10 runs`


Without changes:

`hyperfine --warmup 30 'cat ~/urls | ./fff -o ./mydir'
Benchmark 1: cat ~/urls | ./fff -o ./mydir
  Time (mean ± σ):      2.017 s ±  0.034 s    [User: 0.041 s, System: 0.031 s]
  Range (min … max):    1.980 s …  2.072 s    10 runs`

Content of urls:

`https://www.google.de
https://www.facebook.com
https://www.apple.com
https://www.google.de
https://www.google.de
https://www.google.de
https://www.google.de
https://www.google.de
https://www.google.de
https://www.google.de
https://www.facebook.com
https://www.facebook.com
https://www.facebook.com
https://www.facebook.com
https://www.facebook.com
https://www.facebook.com
https://www.facebook.com`